Update patch and build Makefile commands to handle more args
Currently, cargo make patch (a subtask) assumes all cargo make
flags are repos to patch, which is not the case. This updates it
to expect --crate-patch <repo> and process that.

At the same time, the build tasks are updated to look for
--features <feature_string> to add to the build so that features
can be added to the build via CLI, e.g. --features enable_debugger

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>