Fix Crate Patch on Linux
The crate patch duckscript was assuming that all paths returned by
cargo-metadata were getting `path+file:///` prepended and so
stripping it to get to the real path. However, this is only true
on Windows. On Linux, only `path+file://` is prepended. When the
extra slash is removed on Linux, the patch fails because now the
root dir is removed and the repo can't be found.

e.g. Windows paths come in looking like `path+file:///d:/r/patina`
and Linux paths come in looking like `path+file:///workspaces/patina`.
Note that Linux does have three slashes, but the last one is actually
part of the path (the repo living at /workspaces/patina).

This updates the duckscript to strip by OS accordingly.